load games from gam files. tests for lcp hoswon lover on polymatrix games.
