+++ quantecon/game_theory/game_converters.py
@@ -1,0 +1,33 @@
+from .normal_form_game import NormalFormGame
+from itertools import product
+
+def qe_nfg_from_gam_file(filename):
+      
+    with open(filename, 'r') as file:
+        lines = file.readlines()
+        combined = [
+            token
+            for line in lines
+            for token in line.split()
+        ]
+
+        i = iter(combined)
+        players = int(next(i))
+        actions = [int(next(i)) for _ in range(players)]
+        
+        nfg = NormalFormGame(actions)
+
+        entries = [
+                {
+                    tuple(reversed(action_combination)): float(next(i))
+                    for action_combination in product(*[range(a) for a in actions])
+                }
+                for _ in range(players)
+            ]
+
+        for action_combination in product(*[range(a) for a in actions]):
+            nfg[tuple(reversed(action_combination))] = [
+                entries[p][action_combination] for p in range(players)
+                ]
+
+    return nfg
